Fontend makeover and fixed product deletion

+++ backend/controllers/products.js
@@ -52,6 +52,19 @@ productsRouter.post('/', async (request, response, next) => {
 })
 
 productsRouter.delete('/:id', async (request, response, next) => {
+  const token = getToken(request)
+  const decodedToken = jwt.verify(token, SECRET)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  user.products = user.products.filter(product => product.toString() !== request.params.id)
+  await user.save()
+
   await Product.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
